+++ app/page.tsx
@@ -1,0 +1,42 @@
+import Head from 'next/head'
+
+import { compareByDay, clampDay } from '../utils/date'
+import { load, deserialize } from '../utils/ical'
+import type { Event } from '../utils/ical'
+import { Main } from '../components/Main'
+import '../styles/globals.css'
+
+const calculateSomeStateThings = (events: Event[], currentDay: Date) => {
+  const sortedDates = events.flatMap((event) => [event.start, event.end]).sort(compareByDay)
+  const startingDay = sortedDates[0]
+  const endingDay = sortedDates[sortedDates.length - 1]
+  const initialDay = clampDay(startingDay, endingDay, currentDay)
+
+  return {
+    startingDay,
+    endingDay,
+    initialDay
+  }
+}
+
+const Home = async () => {
+  if (process.env.ICAL_URL === undefined) throw new Error('ICAL_URL is not defined')
+  const serializedEvents = await load(process.env.ICAL_URL)
+  const events = deserialize(serializedEvents)
+
+  const { startingDay, endingDay, initialDay } = calculateSomeStateThings(events, new Date())
+
+  return (
+    <div className="h-screen w-screen">
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <Main startingDay={startingDay} endingDay={endingDay} initialDay={initialDay} events={events} />
+    </div>
+  )
+}
+
+export default Home
