--- pages/index.tsx
@@ -1,69 +1,0 @@
-import { useState } from 'react'
-import Head from 'next/head'
-import type { NextPage } from 'next'
-
-import Calendar from '../components/Calendar'
-import Header from '../components/Header'
-import { isEarlierDay, compareByDay, clampDay } from '../utils/date'
-import { load, deserialize } from '../utils/ical'
-import type { Event, SerializedEvent } from '../utils/ical'
-
-export const getStaticProps = async () => {
-  if (process.env.ICAL_URL === undefined) throw new Error('ICAL_URL is not defined')
-  const serializedEvents = await load(process.env.ICAL_URL)
-  return {
-    props: {
-      serializedEvents
-    }
-  }
-}
-
-const calculateSomeStateThings = (events: Event[], currentDay: Date) => {
-  const sortedDates = events.flatMap((event) => [event.start, event.end]).sort(compareByDay)
-  const startingDay = sortedDates[0]
-  const endingDay = sortedDates[sortedDates.length - 1]
-  const initialDay = clampDay(startingDay, endingDay, currentDay)
-
-  return {
-    startingDay,
-    endingDay,
-    initialDay
-  }
-}
-
-const Home: NextPage<{ serializedEvents: SerializedEvent[] }> = ({ serializedEvents }) => {
-  const events = deserialize(serializedEvents)
-
-  const { startingDay, endingDay, initialDay } = calculateSomeStateThings(events, new Date())
-  const [date, setDate] = useState(initialDay)
-
-  const canPrev = isEarlierDay(startingDay, date) // startingDate < date
-  const canNext = isEarlierDay(date, endingDay) // date < endingDate
-
-  const prev = () => {
-    const newDate = new Date(date)
-    newDate.setDate(date.getDate() - 1)
-    setDate(newDate)
-  }
-
-  const next = () => {
-    const newDate = new Date(date)
-    newDate.setDate(date.getDate() + 1)
-    setDate(newDate)
-  }
-
-  return (
-    <div className="h-screen w-screen">
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <Header date={date} onPrev={prev} onNext={next} canPrev={canPrev} canNext={canNext} />
-      <Calendar events={events} date={date} />
-    </div>
-  )
-}
-
-export default Home
